Add and apply .clang-format
Clang-format, so far as version 3.8 at least, doesn't seem to support my
preferred brace wrapping style for lambdas. To match functions, I prefer
the brace to be wrapped onto the next line. Until this is possible with
clang-format I will submit to the style in this commit with the opening
braces on the same line.